الجزء الأول في المجموعة Phase M

+++ FinalLabSystem/Models/TestResult.cs
@@ -14,7 +14,7 @@ public partial class TestResult
 
     public string? ResultValue { get; set; }
 
-    public double? ResultNumeric { get; set; }
+    public decimal? ResultNumeric { get; set; }
 
     public string? ResultStatus { get; set; }
